[None][fix] address Gemma4 PR review feedback
